Fix translation bug with formatted strings
Try translation with detail level added without any arguments to
determine if the key with the detail level exists.

Translation based on detail level was failing with formatted strings
because it determines the translation does not exist if the string returned by
i18n.Tr is the same as the key sent. This is not the case if additional
arguments are sent.